Refactor the database interface into its own directory.
* It is nicer to separate the API from the database connections.
* Some of the tables don't even have API endpoints anymore - since changing the front end table.
