+++ admin/database_connection/main.php
@@ -1,0 +1,8 @@
+<?php
+
+require_once 'base.php';
+require_once 'cf7_posts.php';
+require_once 'odoo_connections.php';
+require_once 'odoo_errors.php';
+require_once 'odoo_form_mappings.php';
+require_once 'odoo_forms.php';
