+++ admin/database_connection/odoo_errors.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace odoo_conn\admin\api\endpoints;
+namespace odoo_conn\admin\database_connection;
 
 
 trait OdooConnOdooErrorsTableName
