+++ admin/pages/errors/odoo_errors.php
@@ -2,10 +2,8 @@
 
 namespace odoo_conn\admin\pages\errors;
 
-require_once(__DIR__ . "/../../api/endpoints/odoo_errors.php");
-
-use odoo_conn\admin\api\endpoints\OdooConnGetOdooErrors;
-use odoo_conn\admin\api\endpoints\OdooConnDeleteOdooErrors;
+use odoo_conn\admin\database_connection\OdooConnGetOdooErrors;
+use odoo_conn\admin\database_connection\OdooConnDeleteOdooErrors;
 use odoo_conn\admin\pages\OdooConnPageRouter;
 use odoo_conn\admin\table_display\OdooConnCustomTableDeletableDisplay;
