+++ admin/pages/forms/odoo_form.php
@@ -2,13 +2,13 @@
 
 namespace odoo_conn\admin\pages\forms;
 
-require_once(__DIR__ . "/../../api/endpoints/odoo_forms.php");
+require_once __DIR__ . '/../../api/endpoints/odoo_forms.php';
 
-use odoo_conn\admin\api\endpoints\OdooConnGetOdooForm;
-use odoo_conn\admin\api\endpoints\OdooConnDeleteOdooForm;
-use odoo_conn\admin\api\endpoints\OdooConnPostOdooForm;
-use odoo_conn\admin\api\endpoints\OdooConnGetOdooFormSingle;
-use odoo_conn\admin\api\endpoints\OdooConnPutOdooForm;
+use odoo_conn\admin\database_connection\OdooConnGetOdooForm;
+use odoo_conn\admin\database_connection\OdooConnDeleteOdooForm;
+use odoo_conn\admin\database_connection\OdooConnPostOdooForm;
+use odoo_conn\admin\database_connection\OdooConnGetOdooFormSingle;
+use odoo_conn\admin\database_connection\OdooConnPutOdooForm;
 use odoo_conn\admin\pages\OdooConnPageRouterCreate;
 use odoo_conn\admin\table_display\OdooConnCustomTableEditableDisplay;
